interfaces 2.0

+++ IT.Hashing/IHashAlgorithm64.cs
@@ -2,5 +2,5 @@
 
 public interface IHashAlgorithm64 : IHashAlgorithm
 {
-    uint GetHash64();
+    ulong GetHash64();
 }
